moved example
