+++ src/homeworks/homework16/app/game/Dice.java
@@ -1,0 +1,5 @@
+package homeworks.homework16.app.game;
+
+public interface Dice {
+    int roll();
+}
