+++ src/homeworks/homework16/app/game/DiceImpl.java
@@ -1,0 +1,11 @@
+package homeworks.homework16.app.game;
+
+import java.util.Random;
+
+public class DiceImpl implements Dice {
+    @Override
+    public int roll() {
+        // https://docs.oracle.com/javase/8/docs/api/java/util/Random.html
+        return new Random().nextInt();
+    }
+}
