+++ src/homeworks/homework16/app/game/GameWinnerConsolePrinter.java
@@ -1,0 +1,8 @@
+package homeworks.homework16.app.game;
+
+public class GameWinnerConsolePrinter implements GameWinnerPrinter {
+    @Override
+    public void printWinner(Player winner) {
+        System.out.printf("Победитель: %s%n", winner.getName());
+    }
+}
