+++ src/homeworks/homework16/app/game/GameWinnerPrinter.java
@@ -1,0 +1,5 @@
+package homeworks.homework16.app.game;
+
+public interface GameWinnerPrinter {
+    void printWinner(Player winner);
+}
