+++ src/homeworks/homework16/app/game/Player.java
@@ -1,0 +1,13 @@
+package homeworks.homework16.app.game;
+
+public class Player {
+    private final String name;
+
+    public Player(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
