--- src/homeworks/homework16/game/Dice.java
@@ -1,5 +1,0 @@
-package homeworks.homework16.game;
-
-public interface Dice {
-    int roll();
-}
