--- src/homeworks/homework16/game/DiceImpl.java
@@ -1,11 +1,0 @@
-package homeworks.homework16.game;
-
-import java.util.Random;
-
-public class DiceImpl implements Dice {
-    @Override
-    public int roll() {
-        // https://docs.oracle.com/javase/8/docs/api/java/util/Random.html
-        return new Random().nextInt();
-    }
-}
