--- src/homeworks/homework16/game/GameWinnerConsolePrinter.java
@@ -1,8 +1,0 @@
-package homeworks.homework16.game;
-
-public class GameWinnerConsolePrinter implements GameWinnerPrinter {
-    @Override
-    public void printWinner(Player winner) {
-        System.out.printf("Победитель: %s%n", winner.getName());
-    }
-}
