--- src/homeworks/homework16/game/GameWinnerPrinter.java
@@ -1,5 +1,0 @@
-package homeworks.homework16.game;
-
-public interface GameWinnerPrinter {
-    void printWinner(Player winner);
-}
