--- src/homeworks/homework16/game/Player.java
@@ -1,13 +1,0 @@
-package homeworks.homework16.game;
-
-public class Player {
-    private final String name;
-
-    public Player(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
